requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.5
SQLAlchemy~=1.4.27